Bug fixes
The scheduler now schedules things at the right time. The three exam
constraints has been disabled. I *STRONGLY* recommend someone boots up
one of the machines for our use, enables that three exam constraint
checker, and runs it overnight to see if it works.

To enable checking that constraint, uncomment the block in
getSmallestAvailableColour that calls the method.